+++ src/mpid/ch4/netmod/ofi/netmod_inline.h
@@ -11,13 +11,12 @@
 #ifndef NETMOD_INLINE_H_INCLUDED
 #define NETMOD_INLINE_H_INCLUDED
 
+/* Inlined OFI netmod functions */
 #include "ofi_am.h"
 #include "ofi_events.h"
-#include "ofi_comm.h"
 #include "ofi_proc.h"
 #include "ofi_progress.h"
 #include "ofi_unimpl.h"
-#include "ofi_init.h"
 #include "ofi_coll.h"
 #include "ofi_datatype.h"
 #include "ofi_op.h"
@@ -26,7 +25,8 @@
 #include "ofi_send.h"
 #include "ofi_win.h"
 #include "ofi_rma.h"
-#include "ofi_spawn.h"
 #include "ofi_startall.h"
+/* Not-inlined OFI netmod functions */
+#include "ofi_noinline.h"
 
 #endif /* NETMOD_INLINE_H_INCLUDED */
