+++ src/mpid/ch4/netmod/src/Makefile.mk
@@ -1,0 +1,8 @@
+## -*- Mode: Makefile; -*-
+## vim: set ft=automake :
+##
+## (C) 2018 by Argonne National Laboratory.
+##     See COPYRIGHT in top-level directory.
+##
+
+mpi_core_sources += src/mpid/ch4/netmod/src/netmod_impl.c
